python_flask_helm_flux/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
sqlalchemy==2.0.40
PyYAML==6.0.2
gunicorn==23.0.0
jinja2==3.1.6
pytest==8.2.2
pylint==3.3.7
pytest-cov==5.0.0
pytest-html==4.1.1
pydantic==2.10.6
pydantic_core==2.27.2
beautifulsoup4==4.13.4
